refactor : useCategory
카테고리 관련 코드 리팩토링

Related to : #163
